move suite setup to separate file
